Update to the page tracking delay
Cleans up a feature we just merged that delays the page call.  More info below but the short of it is, there's something funky going on with gatsby router that causes it to be out of sync with the actual page, thereby recording the page title incorrectly (it records it as the previous page title).

PR: https://github.com/benjaminhoffman/gatsby-plugin-segment-js/commit/b2a536ad77b9f4324d4dbee030fed9334a9b5fe6